# Conflicts:
#	Source/.env
#	Source/routes/web.php
#	Source/vendor/composer/autoload_classmap.php
#	Source/vendor/composer/autoload_static.php

+++ Source/app/Http/Controllers/BoMonController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Crypt;
 use App\BoMon;
 
 class BoMonController extends Controller
@@ -22,12 +23,12 @@ class BoMonController extends Controller
     {
         $this->validate($request,
             [
-                'TenBM'=>'required|min:3|max:50'
+                'TenBM'=>'required|min:5|max:255'
             ],
             [
                 'TenBM.required'=>'Bạn chưa nhập tên bộ môn',
-                'TenBM.min'=>'Tên bộ môn có ít nhất 3 ký tự',
-                'TenBM.max'=>'Tên bộ môn có nhiều nhất 50 ký tự'
+                'TenBM.min'=>'Tên bộ môn có ít nhất 5 ký tự và có nhiều nhất 255 ký tự',
+                'TenBM.max'=>'Tên bộ môn có ít nhất 5 ký tự và có nhiều nhất 255 ký tự'
             ]);
         
         $bomon = new BoMon;
@@ -39,7 +40,9 @@ class BoMonController extends Controller
 
     public function getSua($id)
     {
+        $id = Crypt::decrypt($id);
         $bomon = BoMon::find($id);
+        if(is_null($bomon)) return redirect('admin/bomon/danhsach');
         return view('admin.bomon.sua', ['bomon'=>$bomon]);
     }
 
@@ -55,19 +58,17 @@ class BoMonController extends Controller
                 'TenBM.min'=>'Tên bộ môn có ít nhất 3 ký tự',
                 'TenBM.max'=>'Tên bộ môn có nhiều nhất 50 ký tự'
             ]);
-        
-        $bomon = new BoMon;
         $bomon->TenBM =$request->TenBM;
-        $bomon->save();        
-
+        $bomon->save();
         return redirect('admin/bomon/sua/'.$id)->with('thongbao','Sửa thành công');
     }
 
     public function getXoa($id)
     {
+        $id = Crypt::decrypt($id);
         $bomon = BoMon::find($id);
+        if(is_null($bomon)) return redirect('admin/bomon/danhsach');
         $bomon->delete();
-
         return redirect('admin/bomon/danhsach')->with('thongbao','Xóa thành công');
     }
 }
